requirements.txt
dwave-ocean-sdk>=3.3.0
pandas==0.25.3

matplotlib==3.1.3; python_version>'3.5'
matplotlib==3.0.3; python_version=='3.5'
